Remove log statements in _on_received_items to fix perf issues.
If a game has a large amount of items and receives them all in a single command (such as when connecting to the server, or when a game is released or collected), the overhead of all the log messages was large enough to cause a several second freeze in-game.

Just remove the messages and leave a warning for the future to not keep them if they're necessary. This part of the code should be stable now anways.