Release v0.0.6 + bugfixes + changelog fixes (#11)
* Fill in Brotato apworld game page

* Add AP error icon to .import

* Remove deprecated default argument from mod_main._init.

There was a warning in ModLoader in the released Brotato about this. It probably happened in a recent update.

* Change AP connect defaults to release values

* Refer to enums in other files in a release-friendly way.

The release version of Brotato can't automatically detect types in our mods if they aren't declared in the same file they're used in.

To fix, we preload the relevant namespace and use the fully qualified name for the enum.

* Remove type annotation that was causing issues

* Remove/fix type references, add more "if connected" guards.

* Set new version in changelog

* Add release dates and diff URLs to changelog

This makes it match the Keep a Changelog format more closely, and adds better history tracking.