Add workaround to set number of wins needed in certain cases.
This fixes the issue where the game could be generated and require more wins than characters available if the player does not have the DLC. This checks if the DLC is not installed, and if so sees if the bug is present. If so, it adjusts the number of wins to a possible value.