Fix wins required bug and incorrect character categorization, with mitigation (#52)
* Get non-loot-crate tests pasing with character pack changes

* Fix num_victories able to be larger than the number of characters

* Recategorize 1.1.0.0 characters as base game chars

I mistakenly put Baby, Vagabond, Technomage, and Vampire in the DLC characters group, despite them actually being part of the base game after 1.1.0.0. This should fix that.

* Add workaround to set number of wins needed in certain cases.

This fixes the issue where the game could be generated and require more wins than characters available if the player does not have the DLC. This checks if the DLC is not installed, and if so sees if the bug is present. If so, it adjusts the number of wins to a possible value.

* Remove test code

* Update changelog

* Remove debug code (good job past me for the comment)

* Fix typo in comment

* Update changelog (again)