Delay before calling notify_run_started to avoid game crash
If we call notify_run_started, and there's enough banked XP to level up, the game will crash because it tries to show the level up text before the floating text Scene (specifically, its Tween) is initialized.

This seems to be an implementation detail change in the order of setup operations for Brotato version 1.1.8.0, which updated the game to use Godot 3.6.

The fix added is to just wait 10ms before sending the notification and continuing through main._ready to let the rest of the game finish setting up. Not pretty, but it handles the problem well enough, and I couldn't see a better way to do it.