Fix logic for determining the bottom neighbor of the AP connect button
The data structure for ProgressData.saved_run_state changed, and our previous check was not valid, so we aloways set the bottom_neighbor of
the AP connect button to the continue button, even if it wasn't visible.

The new check properly determines if there is a saved run or not.