Fix too many pickups spawning (#72)
A recent update changed the base `spawn_consumables` function to have an
`await` inside it to delay until the item was spawned before finishing.
For whatever reason, I couldn't chain the await into the mod code, so I
decided to just move all of the logic into `item_service` instead. This
has the benefit of consolidating the logic into one place.

This also necessitated moving the debug crate spawning logic to a new
location, which wound up being a split of the entity `die` method, and
`item_service`.