Add item weights for items, upgrades, XP and gold (#74)
This adds several new options to set the weight of the aforementioned
items to control how often they appear in the item pool.

The main motivation for this was to fix how these items were generated
and added to the pool. Using weights, we can now simply calculate the
number of locations in the world based on the options, determine from
that how many items we need in addition to characters/shop related
items, and generate the rest using the weights. This is a much more
simple and reliable system than creating a requested amount of items,
then reducing the counts until they match the number of locations.

Using weights also matches how existing apworlds implement similar
features. It's also more "honest" to users, since in the past we were
treating the number of upgrades as upper bounds instead of hard limits.

This also removes the upgrade count options, and the item weight option,
since they aren't relevant with the weights.

One change also hidden in here is that legendary loot crates are now
marked as normal locations instead of excluded, as a bandage to solve
fill error issues. Not the best place for it, but we're making big
changes in the test infra next so w/e.