Reorganize code into units to make testing easier and more robust, and fix subTest bug (#75)
The tests were getting hard to manage because they had to go through the
entire world generation for each step, then check some output property
on the world (items, regions, etc.) for validation. This was slow and
clunky, and easy to make the tests fail due to some other logic.

Instead, I'm moving a lot of logic out of the world class back into
separate modules. All of this was separate originally, and consolidated
into the world in a previous PR to prevent having to send various bits
of state to different functions. But, the benefits of being more
testable outweigh the downside of verbose function calls. In a way, the
extra arguments in the functions are handy to see what options each part
of the world generation relies on.

This essentially turned into a refactor of the code that doesn't have
any actual logic changes. The code is more modular, and the tests are
smaller and can run faster, since they now use `TestCase` in a lot of
places instead of `WorldTestBase` via `BrotatoTestBase`.

The tests also have been fixed so they reliably run now, regardless of
whether they are run individually or as part of a larger test run. This
was not the case previously, since there was a bug in
`data_test_subtest` that caused the options dictionary of the
`BrotatoTestBase` class to get overwritten, leading to unreliable test
results.