Fix item wave values to properly distribute them over waves 1-20 (#77)
The attempt to bias the items towards lower waves just meant that the
game wouldn't give any items higher wave numbers if the number of items
was less than 20.

It's probably not great for balance, but there's bigger issues in that
regard (this barely does anything in-game), and it leads to more natural
results. For example, there will always be an item of wave 20 now, and
the waves will actually be spaced along `[1, 20]` using a cheap
`linspace` implementation.