Add support for AP-world per-character saves and wave skipping (#87)
Add the ability to save runs for each character while connected to a
multiworld, and to resume them later.
Add the option to skip to any wave that the player previously reached
with the current character. This is available from the shop below the
"Go" button.

The big changes for this involved adding the new UI to show the saved
run state and the option to resume it or start a new run, and adding all
the relevant data storage options to track saves and the last character
played.

There are no apworld changes apart from adding a QoL section to the
world description.