Fix game disconnect from large saves and other network-related errors (#90)
The saved runs could be larger than the client's incoming buffer size of
256 KB, which would cause the data storage with the saves to be dropped
by the client when connecting to the server, which ultimately would lead
to the server disconnecting the client. This ups the in- and out-going
buffer sizes to 10 MB. It's a bit overkill, but this isn't the first
time buffer issues have caused bugs like this, and I want this to be the
last.

We also did some smaller changes along this:
* Better logging in the WebSocket client to help detect issues faster.
* The client now tries to connect over `wss://` first, then tries
`ws://` if that fails, rather than the opposite.
* This makes connecting to archipelago.gg (which I am assuming is the
most common connection) a bit faster.
* Update the `Connect` message version from 0.5.0 to 0.6.2.