Compress saved runs and increase WS buffer (#93)
Apparently the 10 MB buffer size was not large enough if enough large
saved runs were stored in the AP data storage. Increase the buffer size
to 20 MB and see if that's enough.

To avoid this problem in the future, saved runs are now compressed using
Zstandard before being sent to data storage. This greatly decreases the
save size, and long-term should hopefully mean saves all comfortably fit
in a smaller buffer.

Old, uncompressed saves are still supported, and will be automatically
compressed the next time they are saved.