Fix run won on death (#94)
The `_is_run_lost` flag is not set until the base game finishes its end
of wave cleanup, and we were checking the flag before this happens. So,
our wins tracker was erroneously giving a win if the player died on wave
20.

Reorder the sequence of base/mod clean up to fix this.