Fix errors when retrying and loading runs (#96)
When a wave is failed, the mod calls `game_state.notify_run_finished` to
indicate we're no longer in a run. However, if the player uses the
option to restart the run from the last wave, we weren't calling
`notify_run_started` properly. This is now fixed, and the check for
calling `notify_run_started` in our `main` extension now automatically
calls the method if we're connected to an AP server and it hasn't been
called already by the resume run flow.

When retrying a wave, the game just rolls back `RunData` to its value at
the start of the wave, so any setup we did for the run is already
applied. To make sure that we don't redo these steps, we add a new
argument to `notify_run_started` and the `game_state.run_started`
signal, `is_new_run`, which is only set by the `main` extension if it
notices that there's one more retries listed in `RunData`.

Also, fixed `ApSavedRunsProgress.get_last_saved_run` not decompressing
the saved run data first, causing issues when trying to resume a run.