QOL changes to the character selection menu (#100)
Updated the character progress handler and the character selection UI to
make it more obvious what characters are included in the player's slot,
and which ones are locked. The exact list of changes:

* The character progress now determines which characters are available
by checking if their "Run Won" location is in the location pool.
* This method means we can determine the available characters for
existing games.
* The character selection menu only shows these available characters,
instead of the entire roster.
* Locked characters (i.e., characters who the player hasn't received the
corresponding AP item for yet), now show their character portrait
instead of the padlock icon to make it clearer in the UI which
characters they have.
* This is particularly useful when the player generates the slot with a
random selection of characters.
* Locked characters have a slightly translucent image (just like the
vanilla padlock icon) to make it clear they're not unlocked yet.
* The lock panel in the character selection menu now shows the text
"Missing AP item for <character name>" instead of their default message.