Fix non deathlink Bounced disconnecting game (#110)
The DeathLink handler was checking for the `tags` field on every
`Bounced` message, but the messages only have the field if it's
non-empty. This was leading to errors in the mod causing the connection
to be dropped. To players, it appeared as if the game would disconnect a
short while after connecting.

This is fixed by checking for the `tags` field before accessing it.

As part of these changes, the tag handling was also updated to fix the
room's tags being applied to the player, and to allow other tags to be
set on the slot. The `enable_deathlink` method is now just a wrapper
around the generic `set_tag` method.