Remove AP required version from template YAML (#111)
Adds a tiny Python script which can remove the `required->version` key
from Archipelago options YAML files. This then hooks up the CD process
to run the script on the template YAML, `Brotato.yaml`, created and
bundled with the release.

Since we make releases from the HEAD of the Archipelago repo, the
version in the template winds up being whatever the in-development
version is, which is typically newer than the latest stable release.
This causes generation errors when used with non-source AP.

I turned this into a CD process because I've consistently forgotten to
do this one too many times :).